[analyzer] Remove deprecated class `BaseDeprecatedMemberUseVerifier`.
This class was left in
https://dart-review.googlesource.com/c/sdk/+/450263 as a temporary
measure since it was used in internal Google sources. Now that those
internal sources have been migrated to use `MemberUseVerifier`
instead, it is no longer needed.

Removing it also allows the class `_TransitionalMemberUseReporter` to
be removed, as well as the `element` parameter of
`MemberUseReporter.report`.

Change-Id: I6a6a69644965549c879e4d25f75e0db7c97af090
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/450362
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>