[messages] Remove unused toYaml methods.
These methods were useful when I was first adding code generation
logic for analyzer error codes, and so I was creating yaml files to
act as the source for code generation. Now that the code generation
logic is fully in place, there is no need for them anymore, and
maintaining them was proving to be a pain.

Change-Id: I6a6a69648498a4994fac4d814153b6500ec2978d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451823
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>