[messages] Remove _SyntacticErrorGenerator.checkForManualChanges.
Removes the method `_SyntacticErrorGenerator.checkForManualChanges`,
which was part of a previous effort to unify analyzer and CFE error
codes, and is not needed by the current effort. Removing this method
will make it easier to clean up and refactor the logic for handling
shared error codes.

Change-Id: I6a6a6964721c4ad0d4ca7336a97346681d476781
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451825
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>