[messages] Remove messages_test.dart.
Removes `pkg/analyzer/tool/messages/messages_test.dart`, which was
useful in the past to verify in the documentation for an analyzer
diagnostic message, the list of parameter descriptions matched the
number of parameters used by the diagnostic message templates.

This test is no longer needed because analyzer diagnostic messages now
specify their parameters in a structured way, and the code generator
checks that the set of parameters used by the diagnostic message
templates matches them. Removing this test will make it easier to
clean up and refactor the logic for handling shared error codes.

Change-Id: I6a6a696400978ce772d4c120733ca19d8ab97f2d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451826
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>