[linter/analyzer_public_api] Report experimental inconsistencies.
Modifies the `analyzer_public_api` lint so that it reports
`analyzer_public_api_experimental_inconsistency` if an element in the
analyzer public API that is _not_ marked experimental has a type that
refers to an element in the analyzer public API that _is_ marked
experimental.

Such references are problematic because they can cause a client of the
analyzer public API to make use of an experimental type without
receiving a warning.

Change-Id: I6a6a6964fd5f1d37e163dc27a1cb4e605efd2d76
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451550
Auto-Submit: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>