[dart2wasm] Simplify handling of name section
Currently the IR has some serialization information attached to it (e.g.
it keeps track of the number of types that have names, etc).

This is redundant/derived information as it can be computed by looking
at the actual IR (e.g. iterate types and count the number of them that
have names).

=> We get rid of this impurity by making the serializer compute the
information it needs to serialize.

I suspect it has been done as an optimization, to avoid an extra pass
over the IR data structures. But this can be handled in another way
witout extra passes as well:

* a single pass traverses and writes the (index, name) pairs
* it keeps track of the number of (index, name) pairs written
* if the number is > 0 then we add the section, the count and the pairs

We do have to traverse all the functions, types, ..., at least once now,
but that's fine: Firstly we always assign names so in reality we cannot
skip these traversals (we instruct binaryen to emit or skip the names
when it runs). Secondly if we wanted to have core dart2wasm not emit
name section we can just skip serializing it if e.g. a
`--no-name-section` flag was passed.

We do a few more changes in this CL:

  * We make `ir.DefType.name` be an optional string
  => This aligns it with `ir.Global.globalName`, `ir.BaseFunction.name`,
     ...

  * Only assign `localNames[local] = ...` and `fieldNames[field] ` ...`
    if name is not empty string

  * We put the `sourceMapUrl` section behind the name section (to align
    with the order used in binaryen)

  * We remove redundancy in the section serializer: Instead of having
    `isEmpty` that has to be kept in sync with `serializeContents` we
    only have `serializeContents` and make the caller check if the bytes
    are non empty.

Issue https://github.com/dart-lang/sdk/issues/60928

Change-Id: Ie9d153fefc25e8277fb30b83e45d2549731797c4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/452040
Reviewed-by: Ömer Ağacan <omersa@google.com>