[messages] Consistently use AnalyzerCode to refer to analyzer error codes.
Creates a new class `AnalyzerCode`, representing an analyzer error
code as a class name and an error name, and consistently uses it in
places that previously represented analyzer error codes as simple
strings.

This change allows for better validation. In follow-up CLs, I will use
this as a vehicle for cleaning up the `messages.yaml` files to refer
to error codes more consistently. For instance:

- Currently, the class name is not specified in all circumstances; I
  will change this so that it is always specified. This in turn will
  allow error codes of type other than `ParserErrorCode` to be shared
  between the analyzer and CFE.

- Currently, some parts of `messages.yaml` refer to errors using
  `UPPER_SNAKE_CASE` while others use `lower_snake_case`. I will
  change this so that `lower_snake_case` is used consistently.

Change-Id: I6a6a696405de7d7e89b1ee11baa776ee639be8fd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/452460
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>