[analyzer] Fix infinite loop when a throw happens when working on a `getIndex` operation
In investigating other things I've gotten into a state where it threw
somewhere inside (something called from) `LibraryAnalyzer.analyze`
making it go into the `} catch (exception, stackTrace) {` in
`AnalysisDriver._analyzeFileImpl`.

In the crashing code path it didn't clear entries in all of the same
maps as in the non-crashing part, namely it forgot the
`_indexRequestedFiles` map, making it - in my case - just again and
again do the getIndex request in an infinite loop.

This CL adds the clearing of this map in the same way as the other maps.

Change-Id: I81a1357cb24ab0a95505e42d00ac06491df09241
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/452620
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>