[messages] Replace remaining CFE analyzerCode values.
Replaces the remaining `analyzerCode: FOO` entries in
`pkg/_fe_analyzer_shared/messages.yaml` with `psuedoSharedCode: FOO`
entries. A `pseudoSharedCode` entry carries two pieces of information:
it indicates that the error code is "pseudo-shared" (meaning that
there is manually maintained logic that translates it from a CFE
diagnostic to an analyzer diagnostic), and it provides an enum value
that can be used by the manually maintained logic to identify which
analyzer error to translate to.

In the few cases where diagnostic had a `pseudoShared: true`
annotation but no `analyzerCode`, a fresh `pseudoSharedCode` is
introduced.

The diagnostic code generation logic is updated accordingly. The logic
that translates CFE diagnostics to analyzer diagnostics is also
updated.

This change carries two benefits:

- It more accurately reflects reality; calling these codes
  `analyzerCode` was inaccurate because there was no automated process
  to verify that they corresponded to actual analyzer error codes, and
  in fact in many cases they didn't.

- It avoids confusion with the `analyzerCode` values in
  `pkg/_fe_analyzer_shared/messages.yaml`, which _do_ in fact
  correspond to actual analyzer error codes.

Change-Id: I6a6a69640fe4b04da0296086d052be46126a1ebd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/453102
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>