[cfe] Update extension member lookup to pair [] and []=
This updates the `ExtensionBuilder.lookupMemberByName`. It is renamed to `lookupExtensionMemberByName` to clearly signal that it is only for extension members and moved to `ExtensionBuilderImpl` so that it is not mixed into other `DeclarationBuilder`s.

Its semantics is changed so that for lookup of `[]` and `[]=` these are returned together as getable and setable, respectively. This simplifies lookups that need both and ensures that we only need a single lookup per name.

Change-Id: I624236972105f8e0ac13cbe3d6f436f2e40ef88e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/453760
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>