[CFE] Expression compilation: #this doesn't have to be the first definition
In expression compilation, to support evaluating "this" in extensions
and extension types the VM sends us a "#this" parameter. The code
assumed this was given as the first variable in the list of definitions,
but that has turned out to not be the case and not something the VM has
ever promised.

This CL allows the #this to be at any place.

Bug: https://github.com/dart-lang/sdk/issues/61502

TEST=Manual

Change-Id: I8868a2e05704cda57f2835e9f261ade12566bf7c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/451760
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Alexander Markov <alexmarkov@google.com>