[messages] Simplify DocumentationValidator.validate.
The for-loop in `DocumentationValidator.validate` was unnecessary; it
iterated through all error classes, and when it found the one error
class for which the `includeCfeMessages` flag is `true`, it then
executed logic that didn't depend on the error class. It's simpler to
simply execute the logic once directly.

Note that the `includeCfeMessages` flag is a bit of a misnomer; what
it actually means is "this is the error class into which diagnostics
that are shared with the CFE should be generated". I will soon be
removing this flag entirely, as part of my effort to allow multiple
error classes to include diagnostics that are shared with the
CFE. Making this change now advances that effort by removing one of
the uses of the flag.

Change-Id: I6a6a6964008a66380a5651d024f7ed5877bff5da
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/454680
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>