analyzer: elementUsageFrontierDetectors: Consider super constructor parameters corresponding to super-parameters
The primary effect this has is to report super-parameters whose
corresponding super-constructor parameters are `@deprecated`. Also
`@experimental`.

Let's say we have a deprecated constructor parameter, and code which
calls that constructor as a super-constructor (implicitly or
explicitly). If the code has no deprecation warnings, it should be
safe to remove the deprecated constructor. This is not the case
without this additional safeguard.

Change-Id: Iacec920d680dfb5d6ccbe19d3f3f54e24a240a9f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/453441
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>