[messages] Remove raw template strings from Template class.
Removes the fields `Template.problemMessageTemplate` and
`Template.correctionMessageTemplate`, which are no longer used.

In addition to simplifying the `Template` class, this prevents the
front end code from relying on details of the template format, which
will give us the freedom to change the template format in the future
if we want to.

Change-Id: I6a6a696489270c63b623d1618698a96dcf6d1bfa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/454961
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>