[messages] Fix backslash escaping in interpolate.
Fixes `_TemplateCompiler.interpolate` so that it properly handles
backslashes in diagnostic messages.

There is no change in generated code because currently, the
`_TemplateCompiler.interpolate` is only used to format messages that
contain placeholders, and none of these messages happen to contain
backslashes.

Paves the way for a follow-up CL in which I will switch to using
`_TemplateCompiler.interpolate` to format all messages.

Change-Id: I6a6a6964f57888893a175e47692ebcc2e9d9099c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/454962
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>