[messages] Rework template parsing.
In the code generators for analyzer and CFE messages, the logic for
parsing the `problemMessage` and `correctionMessage` templates is
moved into the `ErrorCodeInfo` constructor. Templates are now
represented in memory as a list of `TemplatePart`s, each of which is
either a literal string (`TemplateLiteralPart`) or a parameter
substitution (`TemplateParameterPart`).

All the error checking of templates now occurs in the `ErrorCodeInfo`
constructor, so it is not possible for a code generator to skip error
checking. Previously, the analyzer code generator inadvertently
skipped some error checking, which could result in some confusing
behaviors (e.g., if a parameter name was misspelled, the text `null`
was silently interpolated into the diagnostic message).

Change-Id: I6a6a6964a4c47d90ae643ec9c67a955bdcb916cc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455261
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>