[messages] Move translateErrorToken into analyzer.
Moves the function `translateErrorToken`, and the other code it refers
to, from `package:_fe_analyzer_shared` to `package:analyzer`.

The only use of this code in `package:front_end` is in tests, and it's
ok for `package:front_end` tests to refer to `package:analyzer`, so it
didn't need to be shared. Moving it into `package:analyzer` paves the
way for a follow-up CL in which I plan to move all of the analyzer
scanner error codes into `package:analyzer`.

Change-Id: I6a6a6964a155886a4e8fcd2b22b6da38598ce250
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455262
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>