[messages] Move ScannerErrorCode into analyzer.
Moves the class `ScannerErrorCode`, from `package:_fe_analyzer_shared`
to `package:analyzer`.

The only use of this code in `package:front_end` is in tests, and it's
ok for `package:front_end` tests to refer to `package:analyzer`, so it
didn't need to be shared. Moving it into `package:analyzer` allows for
a cleaner separation between analyzer-specific code and code that is
truly shared between the analyzer and the CFE.

Change-Id: I6a6a69643800d664651cffda3a78059a9bb7756d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455263
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>