[analysis_server] Fix folding ranges when nested ranges end on the same line
When the client only supports folding by-line (VS Code), we truncate ranges that intersect but allow nested ranges as long as they don't.

This code was performing truncation if the outer range ended on the same line as the inner range, but this is perfectly fine and didn't need to truncate.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5750

Change-Id: I05e2307574481a7a53ee01ef76bd8a26bfde09c0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/454920
Commit-Queue: Keerti Parthasarathy <keertip@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>