[analysis_server] Skip _refreshAnalysisRoots on open/close of files unless `onlyAnalyzeProjectsWithOpenFiles` is enabled
No functional change, this just skips some work when opening/closing files if we have explicit analysis roots (`onlyAnalyzeProjectsWithOpenFiles=false`, which is the default). In this case, opening and closing files never changes the analysis roots so we don't need to do any of this work.

Change-Id: I052cca6a9b8cbb4d8f936787e83d453f302159c6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455500
Commit-Queue: Keerti Parthasarathy <keertip@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>