[messages] Use fully qualified constant ref in withArguments.
Modifies the analyzer diagnostic message code generator so that when
generating `withArguments` methods, it uses a fully qualified
reference to the diagnostic constant rather than a bare identifier.

Previously, if one of the named parameters of the `withArguments`
method had the same name as the (camelCase) diagnostic constant, the
parameter would shadow the diagnostic constant, so this would cause a
compile-time error in the genrated code.

Note that although this increases the size of the generated code, it
should still compile to the same machine instructions, so there should
be no performance impact.

Thanks to Brian for discovering this issue!

Change-Id: I6a6a6964c0bd33f06a47a12387a2956b018a8fea
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455467
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>