[messages] Rename TransformSetErrorCode.unexpectedToken.
Renames `TransformSetErrorCode.unexpectedToken` to
`TransformSetErrorCode.unexpectedTransformSetToken`. This avoids an
ambiguity between this message and `ParserErrorCode.unexpectedToken`.

Avoiding ambiguities like these is important, because in many cases
the user only sees the diagnostic name; they don't see the class it's
in. For example, `ignore:` comments just give the diagnostic name, and
the web page https://dart.dev/tools/diagnostics only shows diagnostic
names.

In the future I intend to add an error check to the analyzer
diagnostic code generator, to ensure that there are no ambiguities
like these. This CL is a prerequisite for adding the error check.

Change-Id: I6a6a69644ac250b0add9a5f70ec1fd4839ebdef3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455468
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>