[messages] Generalize some references to ScannerErrorCode.
Changes some references to `ScannerErrorCode` so that they refer to
the base class `DiagnosticCode` instead. This paves the way for some
follow-up CLs in which I intend to change the types of some errors
from `ScannerErrorCode` to `ParserErrorCode`.

In the long term I would like for all references to diagnostic in the
analyzer to refer to `DiagnosticCode`, and to keep classes like
`ScannerErrorCode`, `ParserErrorCode`, etc. solely for namespacing
purposes, since there aren't really any important behavioral
differences among the various diagnostic types. This CL is a step in
that direction.

Change-Id: I6a6a69645b06cded8dd0e4acbb23535c8ea8325a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455264
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>