[messages] Clean up `SharedCode` enum.
The enumerated values in the code-generated `SharedCode` enum are now
named after the camelCase analyzer error code, and sorted
accordingly. This makes the enum look a lot more natural and
Dart-like. (Previously, the enumerated values were capitalized and the
sort order wasn't obvious).

The corresponding `sharedAnalyzerCodes` list (which is code-generated
to match the order of the `SharedCode` enum) is suitably adjusted to
reflect the new sort order, so there is no functional change.

Addresses code review comment
https://dart-review.googlesource.com/c/sdk/+/455466/comment/07c4632c_1c93e509/.

Change-Id: I6a6a696469ed5c8a814e3531b666bb358979cde0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455761
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>