[cfe] Implement static extension methods in the CFE
This CL adds the support for the resolution of static extension
methods in the CFE. For that, a minor reorganization of the code in
`ExpressionGeneratorHelper.resolveAndBuildConstructorInvocation` was
made. Specifically, the returned expression was wrapped into a result
class, so that the unresolved result can be handled outside of
`ExpressionGeneratorHelper.resolveAndBuildConstructorInvocation` and
the extension methods can be looked up instead of reporting the
unresolved error immediately.

Part of https://github.com/dart-lang/sdk/issues/61484
Part of https://github.com/dart-lang/sdk/issues/61486

Change-Id: I20dd071326144dc8e6fa5a9a27208861d3b01ecc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/455620
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>