analyzer: deprecated_member_use_from_same_package: handle super-constructors
This is the sibling CL to https://dart-review.googlesource.com/c/sdk/+/455040. In that CL, I made changes to
BestPracticesVerifier, which handles `deprecated_member_use`, but
neglected to handle this lint rule.

When a generative, non-redirecting constructor does not feature a
super-initiailizer, then the supertype's unnamed constructor is called
implicitly. If that super-constructor is deprecated, then we should
report any implicit call as usage.

Change-Id: I990e7f53da5f5ba98612193941e9607f1d3cbca4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/456180
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>